+++ Packages/ArcBoxAuth/Sources/ArcBoxAuth/Storage/KeychainError.swift
@@ -1,0 +1,6 @@
+import Security
+
+public enum KeychainError: Error, Sendable, Equatable {
+    case unhandledStatus(OSStatus)
+    case corruptedItem
+}